+++ package.json
@@ -123,7 +123,6 @@
     "classnames": "^2.2.6",
     "compression": "^1.7.4",
     "cookie-parser": "^1.4.5",
-    "cors": "^2.8.5",
     "dotenv": "^9.0.2",
     "esm": "^3.2.25",
     "express": "^4.17.1",
